Move validateGameDetails to the games module
