+++ src/games.js
@@ -91,5 +91,67 @@ function parseGamelink(messageText){
     return { gamelinkID: gamelinkID };
 }
 
+function validateGameDetails(league, details) {
+    var result = {
+        valid: true,
+        pairing: null,
+        pairingWasNotFound: false,
+        colorsAreReversed: false,
+        gameIsUnrated: false,
+        timeControlIsIncorrect: false,
+        variantIsIncorrect: false,
+        gameOutsideOfCurrentRound: false,
+        reason: ""
+    };
+    var options = league.options.gamelinks;
+
+    var white = details.players.white.userId;
+    var black = details.players.black.userId;
+
+    var potentialPairings = league.findPairing(white, black);
+    var pairing = result.pairing = _.head(potentialPairings);    
+
+    if(potentialPairings.length === 0) {
+        result.valid = false;
+        result.pairingWasNotFound = true;
+        result.reason = "the pairing was not found.";
+    }else if( !details.clock || ( // no clock - unlimited or coorespondence
+        details.clock && ( //clock
+            !_.isEqual(details.clock.initial, options.clock.initial * 60) || // initial time
+            !_.isEqual(details.clock.increment, options.clock.increment) ) // increment
+        ) 
+    ){
+        //the time control does not match options
+        result.valid = false;
+        result.timeControlIsIncorrect = true;
+        result.reason = "the time control is incorrect.";
+    }else if(potentialPairings.length === 1 && pairing && _.isEqual(_.toLower(pairing.white), black)){
+        result.valid = false;
+        result.colorsAreReversed = true;
+        result.reason =  "the colors are reversed.";
+    }else if(!_.isEqual(details.rated, options.rated)){
+        //the game is not rated correctly
+        result.valid = false;
+        result.gameIsUnrated = true;
+        result.reason = "the game is " + ( options.rated ? "unrated." : "rated." );
+    }else if(!_.isEqual(details.variant, options.variant)){
+        //the variant does not match
+        result.valid = false;
+        result.variantIsIncorrect = true;
+        result.reason = "the variant should be standard.";
+    }else{
+        //the link is too old or too new
+        var extrema = scheduling.getRoundExtrema(options);
+        var game_start = moment.utc(details.timestamp);
+        if(game_start.isBefore(extrema.start) || game_start.isAfter(extrema.end)){
+            result.valid = false;
+            result.gameOutsideOfCurrentRound = true;
+            result.reason = "the game was not played in the current round.";
+        }
+    }
+    return result;
+}
+
 module.exports.parseResult = parseResult;
 module.exports.parseGamelink = parseGamelink;
+module.exports.validateGameDetails = validateGameDetails;
